backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# LangChain
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.19
langchain-mistralai==0.1.10

# Vector Store
faiss-cpu==1.8.0

# LLM / Embeddings
openai==1.35.13
ollama==0.2.1

# Document loaders
pypdf==4.2.0
python-docx==1.1.2
unstructured==0.14.8

# Audio
openai-whisper==20231117

# Dev / Test
pytest==8.2.2
httpx==0.27.0
